Added editing and google login feature

+++ frontend/src/components/VendorDetails.jsx
@@ -1,31 +1,29 @@
 import{ useState, useEffect } from 'react';
-import { useParams, useNavigate } from 'react-router-dom';
-import axios from 'axios'; // Import Axios
+import { useParams, useNavigate, Link } from 'react-router-dom';
+import axios from 'axios'; 
 
 function VendorDetails() {
   const navigate = useNavigate()
+
   const handleDelete = async () => {
-    try{
-        await axios.delete(`http://localhost:3000/bank/${id}`)
-        navigate('/')
-    }catch(error){
-        console.log(error)
-    }
-  }
-  const handleEdit = async () => {
     try {
-        await axios.patch(`http://localhost:3000/bank/${id}`)
-        navigate('/add')
+      const shouldDelete = window.confirm('Are you sure you want to delete this vendor?');
+      if (shouldDelete) {
+        await axios.delete(`http://localhost:3000/bank/${id}`);
+        navigate('/');
+      }
     } catch (error) {
-        console.log(error)
+      console.log(error);
     }
-  }
+  };
+  
+
   const { id } = useParams();
   const [vendor, setVendor] = useState(null);
   const [isLoading, setIsLoading] = useState(true);
   const [error, setError] = useState(null); 
-
   useEffect(() => {
+    console.log(id)
     const fetchVendor = async () => {
       try {
         const response = await axios.get(`http://localhost:3000/bank/${id}`);
@@ -55,14 +53,16 @@ function VendorDetails() {
           <p><strong>City:</strong> {vendor.city}</p>
           <p><strong>Country:</strong> {vendor.country}</p>
           <p><strong>Zipcode:</strong> {vendor.code}</p>
+          <button className='delButton' onClick={handleDelete}>Delete</button>
+                <Link to={`/edit/${id}`}>
+                  <button className='editButton'>Edit</button>
+                </Link>
         </div>
       )}
-        <div>
-            <button onClick={handleDelete}>Delete</button>
-            <button onClick={handleEdit}>Edit</button>
-        </div>
     </div>
   );
 }
 
 export default VendorDetails;
+
+
